Score 클래스 생성
