+++ CardGameApp/CardGameApp/Model/Hand.swift
@@ -9,20 +9,9 @@
 import Foundation
 
 class Hand {
-    enum Score: Int, Equatable {
-        case none = 0
-        case onePair
-        case twoPair
-        case triple
-        case fourCard
-        case straight
-        
-        static func >(lhs: Score, rhs: Score) -> Bool {
-            return lhs.rawValue > rhs.rawValue
-        }
-    }
+    
     private var cards = [Card]()
-    private lazy var score = getScore()
+    private lazy var score = Score(cards: cards)
     private lazy var ranks = Ranks(cards: cards)
     
     func forEach(_ transform: (Card) -> ()) {
@@ -33,55 +22,7 @@ class Hand {
         cards.append(card)
     }
     
-    private func isStraight() -> Bool {
-        let sortedCards = cards.sorted { $0 < $1 }
-        var currentCard = sortedCards[0]
-        var straightCount = 1
-        for card in sortedCards {
-            if currentCard.isNextCard(with: card) {
-                straightCount += 1
-                if straightCount == 5 {
-                    return true
-                }
-            } else {
-                straightCount = 1
-            }
-            currentCard = card
-        }
-        return false
-    }
     
-    private func checkPairs(info: [Card:Int]) -> Score {
-        let pairCount = info.filter { $0.value == 2 }.count
-        
-        switch pairCount {
-        case 2...:
-            return .twoPair
-        case 1:
-            return .onePair
-        default:
-            return .none
-        }
-    }
-    
-    private func getScore() -> Score {
-        let info = cards.reduce(into: [Card:Int]()) {
-            if $0[$1] == nil {
-                $0[$1] = 1
-            } else {
-                $0[$1]! += 1
-            }
-        }
-        
-        if isStraight() {
-            return .straight
-        } else if info.contains(where: { $0.value == 4 }) {
-            return .fourCard
-        } else if info.contains(where: { $0.value == 3 }) {
-            return .triple
-        }
-        return checkPairs(info: info)
-    }
 }
 
 extension Hand: Equatable {
